requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
cachetools
certifi
charset-normalizer
colorama
deprecation
Django
djangorestframework
djangorestframework_simplejwt
drf-yasg
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
gotrue
h11
h2
hpack
httpcore
httplib2
httpx
hyperframe
idna
inflection
iniconfig
multidict
oauthlib
packaging
pillow
pluggy
postgrest
propcache
proto-plus
protobuf
psycopg2
psycopg2-binary
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
PyJWT
pyparsing
pytest
pytest-mock
python-dateutil
python-decouple
python-dotenv
pytz
PyYAML
realtime
requests
requests-oauthlib
rsa
six
sniffio
sqlparse
storage3
StrEnum
supabase
supafunc
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
websockets
yarl
